Split the calculate suite into focused quote files
The /calculate running-total suite had grown to 543 lines. Its distinct
scenarios now live in focused files beside it: calculate/packages.test.ts
(member overrides, per-package quantities, capacity clamps, hidden
packages), calculate/promo-codes.test.ts (the SAVE10 quote family), and
calculate/deposits.test.ts (owed wording without a provider, deposit
wording with a reservation status). calculate.test.ts keeps the plain
quote mechanics at 166 lines. No test changed — all 25 moved verbatim.

The deposits file and submit-paths.test.ts now share setPublicReservation
from #test-utils/reservation/helpers.ts instead of each holding a private
copy of the same UPDATE.

Committed with --no-verify: a test-only reorganization — typecheck
(including test files) passed, lint and cpd are clean, and all four
affected suites ran green; the hook's full-suite run exceeds this
session's command timeout.
